+++ quick_blue/CHANGELOG.md
@@ -1,6 +1,23 @@
+## [0.4.1+1] - 2022.3.22
+
+- Add `manufacturerDataHead` & Refactor `manufacturerData` in BlueScanResult
+
+## [0.4.1] - 2022.3.21
+
+- Add `setLogger`
+
+## [0.4.0+1] - 2022.3.21
+
+- Add API compatibility table to README
+
+## [0.4.0] - 2022.3.21
+
+- Add limited Linux support via quick_blue_linux
+- Workaround empty device name on Windows
+
 ## [0.3.1+3] - 2022.3.20
 
-- Fix missing `deviceId` in `characteristicValue` message on iOS/macOs
+- Fix missing `deviceId` in `characteristicValue` message on iOS/macOS
 
 ## [0.3.1+2] - 2022.3.10