+++ quick_blue_linux/CHANGELOG.md
@@ -1,3 +1,12 @@
-## 0.0.1
+## 0.1.2 - 2022.3.22
 
-* TODO: Describe initial release.
+- Refactor `manufacturerData` to `manufacturerDataHead`
+
+## 0.1.1 - 2022.3.21
+
+- Impl `setLogger`
+
+## 0.1.0 - 2022.3.21
+
+- Impl `isBluetoothAvailable` & `scanResultStream`
+- Impl `startScan` & `stopScan`